implemented report module base

+++ resources/views/customer/index.blade.php
@@ -41,10 +41,12 @@
 				<td>{{ $customer->email }}</td>
 				<td>
 					<a href="{{url ('customer', $customer->id)}}"class="btn btn-xs btn-link">Lihat	 </a>
+					@if(Auth::user()->role == 'admin')
 					| <a href="{{ URL::to('customer/' . $customer->id . '/edit') }}"class="btn btn-xs btn-link">Edit </a> | 
-					{!! Form::model($customer, ['method'=> 'DELETE', 'action' => ['CustomerController@destroy', $customer->id],'class'=>'btn btn-xs btn-link']) !!}
+					{!! Form::model($customer, ['method'=> 'DELETE', 'action' => ['CustomerController@destroy', $customer->id],'class'=>'btn btn-xs btn-link', 'onsubmit' => 'return ConfirmDelete()']) !!}
 					{!! Form::submit('Hapus',['class'=>'btn btn-xs btn-link']) !!}
 					{!! Form::close() !!}
+					@endif
 				</td>
 			</tr>
 
@@ -58,4 +60,19 @@
 </div>
 
 
+@stop
+
+
+@section('custom_javascript')
+
+<script type="text/javascript">
+function ConfirmDelete()
+{
+	var x = confirm("Apa anda yakin akan menghapus data ini?");
+	if (x)
+		return true;
+	else
+		return false;
+}
+</script>
 @stop
